:recycle: [Annexe] Refacto with attributes
